+++ MatchWriter/MatchDataConsumer.cs
@@ -48,10 +48,10 @@ namespace MatchWriter
                 // Get matchDataSetJson from redis
                 if(model.ExpiryDate <= DateTime.Now)
                 {
-                    _logger.LogError($"ExpiryDate has already passed. Aborting. Incoming message: {model.ToString()}");
+                    _logger.LogError($"MatchId [ {model.MatchId} ] ExpiryDate has passed. Aborting");
 
                     producer.PublishMessage(msg);
-                    return ConsumedMessageHandling.ThrowAway;
+                    return ConsumedMessageHandling.Done;
                 }
 
                 using (var scope = _sp.CreateScope())
@@ -66,7 +66,7 @@ namespace MatchWriter
                     //Delete uploaded match from redis
                     await _cache.DeleteMatch(model.RedisKey).ConfigureAwait(false);
 
-                    _logger.LogInformation($"Succesfully handled Match#{model.MatchId}.");
+                    _logger.LogInformation($"Successfully handled MatchId [ {model.MatchId} ]");
 
                     msg.Success = true;
 
@@ -75,30 +75,50 @@ namespace MatchWriter
                     return ConsumedMessageHandling.Done;
                 }
             }
+
+            // If a match was expected in Redis but not found
+            catch (Exception e) when (e is MatchEmptyOrNull)
+            {
+                _logger.LogError(
+                    e,
+                    $"MatchId [ {model.MatchId} ] Could not retreive Match from Redis.");
+
+                await _cache.DeleteMatch(model.RedisKey).ConfigureAwait(false);
+
+                producer.PublishMessage(msg);
+                return ConsumedMessageHandling.Done;
+            }
+
             // If it seems like a temporary failure, resend message
             catch (Exception e) when (e is TimeoutException || e is RedisConnectionException)
             {
-                _logger.LogError(e, $"Match#{model.MatchId} could not be uploaded to database right now. Instructing the message to be resent, assuming this is a temporary failure.");
+                _logger.LogError(e, $"MatchId [ {model.MatchId} ] could not be uploaded to database right now. Instructing the message to be resent, assuming this is a temporary failure.");
 
                 return ConsumedMessageHandling.Resend;
             }
+
             // As of now it seems like MatchWriter has a memory leak, which leads to OutOfMemoryException's being thrown for every message. 
             // This catch-block is here to force a restart if this happens. A better solution would be to fix the memory leak.
             catch (Exception e) when (e is OutOfMemoryException)
             {
-                _logger.LogError(e, $"Match#{model.MatchId} could not be uploaded to database right now.");
+                _logger.LogError(e, $"MatchId [ {model.MatchId} ] could not be uploaded to database right now.");
                 _logger.LogCritical($"Exiting the application to force a restart with cleared RAM.");
                 // Exit with Code 14 (ERROR_OUTOFMEMORY)
                 Environment.Exit(14);
                 return ConsumedMessageHandling.Resend;
             }
+
             // When in doubt or the message itself might be corrupt, throw away
             catch (Exception e)
             {
-                _logger.LogError(e, $"Match#{model.MatchId} could not be uploaded to database. Instructing the message to be thrown away, assuming the message is corrupt.");
+                _logger.LogError(
+                    e,
+                    "Unhandled Exception - " +
+                    $"MatchId [ {model.MatchId} ] could not stored in the Database, " +
+                    "Instructing the message to be thrown away.");
 
                 producer.PublishMessage(msg);
-                return ConsumedMessageHandling.Done;
+                return ConsumedMessageHandling.ThrowAway;
             }
         }
     }