+++ CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Upcoming
+
+## [1.1.0] 2020-04-27
+### Changed
+- Default prefetch count to `1`.
+- Added handling of null or empty Match when querying Redis.
 
 ## [1.0.4] 2020-04-08
 ### Changed